refactor(cli): move init business logic into lib
leaving the project conversion out of it since it's currently only partially supported and too entangled with the template
