+++ packages/cdktf-cli/bin/cmds/helper/init.ts
@@ -6,9 +6,9 @@ import { TerraformLogin } from "../helper/terraform-login";
 
 import * as os from "os";
 import * as path from "path";
-import { sscaff } from "sscaff";
+
 import * as terraformCloudClient from "../helper/terraform-cloud-client";
-import { FUTURE_FLAGS } from "cdktf/lib/features";
+
 import { downloadFile, HttpError } from "../../../lib/util";
 import { logFileName, logger } from "../../../lib/logging";
 import { Errors } from "../../../lib/errors";
@@ -27,13 +27,10 @@ import {
   LANGUAGES,
   config,
 } from "@cdktf/provider-generator";
-import { readPackageJson } from "./utilities";
 import { templates, templatesDir } from "./init-templates";
+import { init, Project } from "../../../lib";
 
 const chalkColour = new chalk.Instance();
-
-const pkg = readPackageJson();
-const constructsVersion = pkg.dependencies.constructs;
 
 export function checkForEmptyDirectory(dir: string) {
   if (
@@ -112,17 +109,13 @@ This means that your Terraform state file will be stored locally on disk in a fi
     }
   }
 
-  const deps: any = await determineDeps(argv.cdktfVersion, argv.dist);
-
-  const futureFlags = Object.entries(FUTURE_FLAGS)
-    .map(([key, value]) => `"${key}": "${value}"`)
-    .join(`,\n`);
-
-  await sscaff(templateInfo.Path, destination, {
-    ...deps,
-    ...projectInfo,
-    futureFlags,
+  await init({
+    cdktfVersion: argv.cdktfVersion,
+    destination,
+    dist: argv.dist,
     projectId,
+    projectInfo,
+    templatePath: templateInfo.Path,
   });
 
   if (argv.fromTerraformProject) {
@@ -208,77 +201,6 @@ function copyLocalModules(
         recursive: true,
       })
     );
-}
-
-async function determineDeps(
-  version: string = pkg.version,
-  dist?: string
-): Promise<Deps> {
-  const pythonVersion = version.replace(/-pre\./g, ".dev");
-
-  if (dist) {
-    const ret = {
-      npm_cdktf: path.resolve(dist, "js", `cdktf@${version}.jsii.tgz`),
-      npm_cdktf_cli: path.resolve(dist, "js", `cdktf-cli-${version}.tgz`),
-      pypi_cdktf: path.resolve(dist, "python", `cdktf-${pythonVersion}.tar.gz`),
-      mvn_cdktf: path.resolve(
-        dist,
-        "java",
-        `com/hashicorp/cdktf/${version}/cdktf-${version}.jar`
-      ),
-      nuget_cdktf: path.resolve(
-        dist,
-        "dotnet",
-        `HashiCorp.Cdktf.${version}.nupkg`
-      ),
-      go_cdktf: path.resolve(dist, "go", `cdktf`),
-    };
-
-    for (const file of Object.values(ret)) {
-      if (!(await fs.pathExists(file))) {
-        throw Errors.Internal(
-          "init",
-          `unable to find ${file} under the "dist" directory (${dist})`,
-          { version }
-        );
-      }
-    }
-
-    const versions = {
-      cdktf_version: version,
-      constructs_version: constructsVersion,
-    };
-
-    return {
-      ...ret,
-      ...versions,
-    };
-  }
-
-  if (version === "0.0.0") {
-    throw Errors.Usage(
-      "init",
-      chalkColour`{redBright cannot use version 0.0.0, use --cdktf-version, --dist or CDKTF_DIST to install from a "dist" directory}`,
-      {}
-    );
-  }
-
-  // determine if we want a specific pinned version or a version range we take
-  // a pinned version if version includes a hyphen which means it is a
-  // pre-release (e.g. "0.12.0-pre.e6834d3"). otherwise, we require a caret
-  // version.
-  const ver = version.includes("-") ? version : `^${version}`;
-
-  return {
-    cdktf_version: version,
-    constructs_version: constructsVersion,
-    npm_cdktf: `cdktf@${ver}`,
-    npm_cdktf_cli: `cdktf-cli@${ver}`,
-    pypi_cdktf: `cdktf~=${pythonVersion}`,
-    mvn_cdktf: version,
-    nuget_cdktf: version,
-    go_cdktf: `v${version}`,
-  };
 }
 
 async function gatherInfo(
@@ -500,24 +422,6 @@ async function findCdkTfJsonDirectory(rootDir: string): Promise<string | null> {
     }
   }
   return null;
-}
-
-interface Deps {
-  npm_cdktf: string;
-  npm_cdktf_cli: string;
-  pypi_cdktf: string;
-  mvn_cdktf: string;
-  nuget_cdktf: string;
-  go_cdktf: string;
-  cdktf_version: string;
-  constructs_version: string;
-}
-
-interface Project {
-  Name: string;
-  Description: string;
-  OrganizationName: string;
-  WorkspaceName: string;
 }
 
 interface Template {
