+++ packages/cdktf-cli/lib/index.ts
@@ -1,2 +1,4 @@
 // This is the programmatic entrypoint that the CLI uses.
 // While this is the closest we have to a programmatic API, please understand that the interfaces in this file are not stable.
+
+export { init, Project, InitArgs } from "./init";
